refactoring schema methods
