+++ src/auth/router.js
@@ -2,37 +2,46 @@
 // Create a GET route for /users that returns a JSON object with all users
 'use strict';
 const express = require('express');
-const users = require('./model/users')
+const users = require('./model/user-schema')
 const basicAuth = require('./middleware/basic')
 const ouath = require('./middleware/oauth')
 const app = express();
-const bearerMiddleware=require('./middleware/bearer')
+const bearerMiddleware = require('./middleware/bearer')
 
 app.use(express.json());
 
 
 // Create a POST route for /signup
 app.post('/signup', (req, res) => {
-    users.save(req.body).then(userRecord => {
-        console.log('----singup before generate token',userRecord);
-        let token = users.generateToken(userRecord);
-        res.status(200).send(token);
-    }).catch(err => res.status(403).send("This user name not availble, Error!!!!!!!!!!"));
+    let newUser = req.body;
+    users.findOne({
+            username: newUser.username
+        })
+        .then(result => {
+            if (!result.length) {
+                let user = new users(newUser)
+                user.save()
+                    .then(user => {
+                        let token = users.generateToken(user);
+                        res.status(201).send(token);
+                    }).catch(err => res.status(403).send("This user name not availble, Error!!!!!!!!!!"));
+
+            }
+        })
 })
 
 app.post('/signin', basicAuth, (req, res) => {
-    console.log('respons>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>');
     res.status(200).send(req.token);
 });
 
 app.get('/users', basicAuth, (req, res) => {
     // list all users 
-    users.list().then(data => {
-
-        res.status(200).send(data);
-    })
+    users.find({})
+        .then(data => {
+            res.status(200).send(data);
+        })
 });
-app.get('/oauth', ouath, (req, res)=> {
+app.get('/oauth', ouath, (req, res) => {
     res.status(200).send(req.token);
 });
